Split server-images.test.ts and server-bulk-email.test.ts, drop biome grandfathering
Both files exceeded biome's 1000-line noExcessiveLinesPerFile ceiling and
were grandfathered in biome.json's override list. Split each into a
test/lib/server-images/ and test/lib/server-bulk-email/ folder of
topic-focused files (all well under 400 lines) with a shared helpers.ts,
then removed both paths from the ignore list per AGENTS.md's rule that a
split file's override entry gets deleted, never added to.

Splitting surfaced jscpd duplication that jscpd couldn't detect inside the
original monoliths (per AGENTS.md's noted side effect) - fixed by
extracting shared helpers (linkStoredImage/expectImageErrorRedirect,
seedMarketingDraftWithUnsubscriber, getPreviewHtml/postPreviewHtml,
createListingWithFile) so the 0% duplication threshold still passes.
