+++ biome.json
@@ -170,9 +170,7 @@
         "test/lib/questions.test.ts",
         "test/lib/server-attendee-form.test.ts",
         "test/lib/server-attendees.test.ts",
-        "test/lib/server-bulk-email.test.ts",
         "test/lib/server-groups.test.ts",
-        "test/lib/server-images.test.ts",
         "test/lib/server-listing-parents.test.ts",
         "test/lib/server-modifiers.test.ts",
         "test/lib/server-parents-booking.test.ts",
